Improve - Improve - Move code for Assignment form to separate component, remove unused props, reduce unnecessary prop passing, fix for delete assignment button condition, fix console warnings
