+++ src/components/molecules/AssignmentCandidatesSection/classes.module.scss
@@ -12,6 +12,13 @@
       &:focus {
         box-shadow: inset 0 0 0 2px $color-error;
       }
+
+      &.disabledIcon {
+        cursor: not-allowed;
+        svg path {
+          fill: $color-background-disabled;
+        }
+      }
     }
 
     thead {
