+++ src/components/organisms/DataTable/DataTable.tsx
@@ -123,6 +123,7 @@ const DataTable = <TData,>(
         pageSize: paginationData ? paginationData.per_page : 10,
       })
     }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [current_page])
 
   useEffect(() => {
