+++ src/i18n/locales/et.json
@@ -345,7 +345,6 @@
     "EXPORT_TM": "Asutuse tõlkemälu eksport",
     "DELETE_TM": "Asutuse tõlkemälu eemaldamine",
     "EDIT_TM_METADATA": "Asutuse tõlkemälu andmete muutmine",
-    "EDIT_TM": "Asutuse tõlkemälu muutmine",
     "VIEW_TM": "Asutuse tõlkemälude vaatamine",
     "CREATE_PROJECT": "Tellimuse lisamine",
     "MANAGE_PROJECT": "Tõlketellimuse korraldamine",
