+++ src/types/privileges.ts
@@ -15,7 +15,6 @@ export enum Privileges {
   ExportTm = 'EXPORT_TM',
   DeleteTm = 'DELETE_TM',
   EditTmMetadata = 'EDIT_TM_METADATA',
-  EditTm = 'EDIT_TM',
   ViewTm = 'VIEW_TM',
   CreateProject = 'CREATE_PROJECT',
   ManageProject = 'MANAGE_PROJECT',
